feat: add new service, service form
